Allow setting the default tag name via `name`
The plugin used to use the commit SHA as a tag name passed to
`docker build`. When not using the DIND daemon, this can present a race
condition in scenarios when performing parallel builds: the last build
to finish will tag with the same commit SHA and stomp on the prior
build.

This allows overriding the tag name and keeps the default behaviour
if an override isn't specified.